Fixed file image provider for web
