+++ lib/src/web/extended_file_image_provider.dart
@@ -1,0 +1,7 @@
+import '../extended_image_provider.dart';
+
+class ExtendedFileImageProvider with ExtendedImageProvider {
+  ExtendedFileImageProvider(this.file);
+
+  final Object file;
+}
